front/core/api/entityView: move default components out to components
