+++ frontend/packages/core/src/api/entityView/EntityPageBuilder.tsx
@@ -1,70 +1,14 @@
-import React, { ComponentType, FC } from 'react';
-import { Route, Redirect, Switch } from 'react-router-dom';
-import List from '@material-ui/core/List';
-import ListItem from '@material-ui/core/ListItem';
-import { AppComponentBuilder, App } from '../app/types';
-import { useEntity, useEntityUri, useEntityConfig } from './EntityContext';
-import EntityLink from '../../components/EntityLink/EntityLink';
+import React, { ComponentType } from 'react';
+import DefaultEntityPage from '../../components/DefaultEntityPage';
+import { App, AppComponentBuilder } from '../app/types';
 import BackstagePlugin from '../plugin/Plugin';
-import { Header } from '../..';
-import {
-  EntityPageNavbarProps,
-  EntityPageHeaderProps,
-  EntityPageProps,
-  EntityPageNavItem,
-  EntityPageView,
-} from './types';
+import { EntityPageNavItem, EntityPageView } from './types';
 
 // type AppComponents = {
 //   EntityPage: ComponentType<EntityPageProps>;
 //   EntityPageNavbar: ComponentType<EntityPageNavbarProps>;
 //   EntityPageHeader: ComponentType<EntityPageHeaderProps>;
 // };
-
-const DefaultEntityPageNavbar: FC<EntityPageNavbarProps> = ({ navItems }) => {
-  const entityUri = useEntityUri();
-
-  return (
-    <List>
-      {navItems.map(({ title, target }) => (
-        <ListItem key={target}>
-          <EntityLink uri={entityUri} subPath={target}>
-            {title}
-          </EntityLink>
-        </ListItem>
-      ))}
-    </List>
-  );
-};
-
-const DefaultEntityPageHeader: FC<EntityPageHeaderProps> = () => {
-  const { id } = useEntity();
-  const config = useEntityConfig();
-  return <Header title={`${config.title} - ${id}`} />;
-};
-
-const DefaultEntityPage: FC<EntityPageProps> = ({ navItems, views }) => {
-  const { kind, id } = useEntity();
-  const basePath = `/entity/${kind}/${id}`;
-
-  return (
-    <div>
-      <DefaultEntityPageHeader />
-      <DefaultEntityPageNavbar navItems={navItems} />
-      <Switch>
-        {views.map(({ path, component }) => (
-          <Route
-            key={path}
-            exact
-            path={`${basePath}${path}`}
-            component={component}
-          />
-        ))}
-        <Redirect from={basePath} to={`${basePath}${views[0].path}`} />
-      </Switch>
-    </div>
-  );
-};
 
 type EntityPageRegistration =
   | {
