+++ frontend/packages/core/src/components/DefaultEntityPage/index.ts
@@ -1,0 +1,1 @@
+export { default } from './DefaultEntityPage';
