+++ frontend/packages/core/src/components/DefaultEntityPageHeader/index.ts
@@ -1,0 +1,1 @@
+export { default } from './DefaultEntityPageHeader';
