+++ frontend/packages/core/src/components/DefaultEntityPageNavbar/index.ts
@@ -1,0 +1,1 @@
+export { default } from './DefaultEntityPageNavbar';
